requirements.txt
coppeliasim_zmqremoteapi_client
numpy
matplotlib
pygame
opencv-contrib-python
# push-to-see
scipy
pyyaml
pycocotools
# pytorch
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.13.1+cu116
torchvision==0.14.1+cu116
torchaudio==0.13.1
